fix: search composable renamed
