+++ packages/ui/client/auto-imports.d.ts
@@ -266,8 +266,8 @@ declare global {
   const useScriptTag: typeof import('@vueuse/core')['useScriptTag']
   const useScroll: typeof import('@vueuse/core')['useScroll']
   const useScrollLock: typeof import('@vueuse/core')['useScrollLock']
-  const useSearch: typeof import('./composables/seach')['useSearch']
-  const useSearchTasks: typeof import('./composables/seach')['useSearchTasks']
+  const useSearch: typeof import('./composables/search')['useSearch']
+  const useSearchTasks: typeof import('./composables/search')['useSearchTasks']
   const useSessionStorage: typeof import('@vueuse/core')['useSessionStorage']
   const useShare: typeof import('@vueuse/core')['useShare']
   const useSlots: typeof import('vue')['useSlots']
